FIX #618: test if -lrt is required for linking
This checks for clock_gettime availability in the clib
and in librt and adds -lrt to LUMIERA_LIBS when the
function is available in librt.

The check is only done for autotools based builds,
scons may need something else.

Other functions from the realtime need to be added on
demand.