add backup of the original (unreviewed) GuiBrainstorming page
Somehow I can't find this original page, but according to the
comment in the source, it should have been asciidoced. I'm filing
this backup from the Pipa-Wiki (MoinMoin), because this original page
still contains a wealth of ideas, even if we won't have
the standing to cope with them right now.