WIP reshape test fixture to get a better call structure
This test setup is intended to emulate the situation
when adding jobs to the scheduler; thus we should use
an implicit sequence as root element.
I.e. we have to treat a wood, not a single tree

Note: test still fails, since we take a copy
of a Node object somewhere inadvertently