no need to use pristine-tar
Explanation: pristine-tar can be used in cases where
upstream has the habit of doing arbitrary changes to the
tarball, which aren't recorded in SCM. Typical example is
to include additional binaries, images, videos.

Since we're "upstream" ourselves, we *do* care for a clean
tree and thus the packaging can just draw a diff againtst
the tagged release state. git-buildpackage is able to
re-create the tarball without further help in these cases.